Moved non main precomp code into sep folder
