+++ PrecompilationCode/ArduinoStrategyPattern.h
