+++ PrecompilationCode/IOutputStrategyPattern.h
